Consolidate UICatalog tests to cut down on time outs
