+++ test/e2e/uicatalog/desired.js
@@ -2,5 +2,6 @@ import getAppPath from 'sample-apps';
 import env from '../helpers/env.js';
 
 module.exports = {
-  app: getAppPath('UICatalog', env.REAL_DEVICE)
+  app: getAppPath('UICatalog', env.REAL_DEVICE),
+  noReset: true
 };
